refactor(BreadcrumbLink): remove the interface to index file
