--- src/components/navigation/BreadcrumbLink/interfaces/BreadcrumbLink.interface.ts
@@ -1,10 +1,0 @@
-import { Typos } from "../types/BreadcrumbLink.typos.type";
-
-export interface IBreadcrumbLinkProps {
-  isActive?: boolean;
-  label: string;
-  path: string;
-  id: string;
-  typo?: Typos;
-  handleClick?: () => void;
-}
