+++ src/components/navigation/BreadcrumbLink/stories/BreadcrumbLink.Default.stories.tsx
@@ -1,6 +1,7 @@
+import { BrowserRouter } from "react-router-dom";
 import { BreadcrumbLink } from "../index";
 import { BreadcrumbLinkController } from "./BreadcrumbLinkController";
-import { BrowserRouter } from "react-router-dom";
+import { IBreadcrumbLinkProps } from "..";
 
 import {
   parameters,
@@ -11,7 +12,6 @@ import {
   typo,
   path,
 } from "./props";
-import { IBreadcrumbLinkProps } from "../interfaces/BreadcrumbLink.interface";
 
 const story = {
   title: "navigation/BreadcrumbLink",
