+++ src/components/navigation/BreadcrumbLink/stories/BreadcrumbLinkController.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { BreadcrumbLink } from "../index";
-import { IBreadcrumbLinkProps } from "../interfaces/BreadcrumbLink.interface";
+import { IBreadcrumbLinkProps } from "..";
 
 const BreadcrumbLinkController = (props: IBreadcrumbLinkProps) => {
   const [isActive, setIsActive] = useState(false);
