+++ src/components/navigation/BreadcrumbLink/styles.ts
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { colors } from "../../../shared/colors/colors";
 import { Link } from "react-router-dom";
-import { IBreadcrumbLinkProps } from "./interfaces/BreadcrumbLink.interface";
+import { IBreadcrumbLinkProps } from "./index";
 
 const StyledContainerLink = styled.li`
   display: inline-block;
